✨ feat: Add role and controller for namespace
This commit adds a new Kubernetes RBAC role in `config/rbac/role.yaml` to provide get, list, and watch access to the "namespaces" resource and get access to the "namespaces/status" resource. It also introduces a new controller for the "Namespace" object in `controllers/namespace_controller.go`. The `SetupWithManager` function is called to set up the controller with the manager. Additionally, the `go.mod` file has been updated to include the `k8s.io/api v0.26.0` package as an indirect dependency, and the `REGISTRY` field in `docker/Makefile` has been changed to `registry.gitlab.com/bitspur/community`.