✨ feat: Add new configuration option for CRD auto deletion
Add a new configuration option for automatic deletion of CRDs. The option is added to the chart's values.yaml file and can be set to true or false. When the value is set to true, the controller will automatically delete the CRD resource when the last object using it is deleted. When the value is set to false, the CRD will not be deleted automatically. The new option is also added to the chart's questions.yaml file as a required variable, along with its label and group. The option is then referenced in the easy-olm-operator deployment template as an environment variable. Finally, the crdref_controller.go file is updated to read the value of this option from the environment variable "CRD_AUTO_DELETE" and use it to decide whether to delete the CRD when the last object is deleted. An observed generation field is also added to the CRDRef status to track changes in the CRDRef object.