✨ feat: Add support for InstallPlan reconciliation
This commit adds support for reconciling `InstallPlan`s in the operator. It includes changes to `controllers/installplan_controller.go`, `config/rbac/role.yaml`, and a new file `util/installPlan.go` that provides a utility function for approving an `InstallPlan`. Additionally, RBAC permissions are added to the namespace controller to allow it to watch and create `OperatorGroups`.