✨ feat: add support for OperatorFramework v1 and v1alpha1
This commit adds support for the OperatorFramework v1 and v1alpha1 APIs, including updating the import statements and adding them to the scheme. It also renames the `start` target in the Makefile to `dev`. Additionally, an error handling condition has been added to the `NamespaceReconciler` to check if operator group already exists before creating a new one.